Add wrapContentsInSlot command
